test(data-flow): distribute red team boundary tests logically
Moves out the monolithic 'red_team_test.dart' suite into their logical homes within 'analyzer_test.dart', 'control_flow_escape_test.dart', and 'signature_synthesizer_test.dart', natively aligned with existing test boundaries.

Also resolves lingering analyzer lints on CI (avoid_dynamic_calls, omit_local_variable_types, and deprecated_member_use).
